feat(migrations): change total_expected column type from INTEGER to BIGINT

+++ migrations/027_total_expected.sql
@@ -1,1 +1,1 @@
-ALTER TABLE pulso.sync_jobs ADD COLUMN IF NOT EXISTS total_expected INTEGER;
+ALTER TABLE pulso.sync_jobs ADD COLUMN IF NOT EXISTS total_expected BIGINT;
